Hw 11

+++ ray_spring/spring-practice/src/main/java/com/example/spring_practice/domain/comment/service/CommentService.java
@@ -10,6 +10,7 @@ import com.example.spring_practice.domain.member.entity.Member;
 import com.example.spring_practice.domain.member.repository.MemberRepository;
 import com.example.spring_practice.domain.post.entity.Post;
 import com.example.spring_practice.domain.post.repository.PostRepository;
+import com.example.spring_practice.domain.shared.ImageService;
 import com.example.spring_practice.global.response.CustomException;
 import com.example.spring_practice.global.response.ErrorCode;
 import lombok.RequiredArgsConstructor;
@@ -22,11 +23,12 @@ import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
+@Transactional(readOnly = true)
 public class CommentService {
     private final CommentRepository commentRepository;
     private final PostRepository postRepository;
     private final MemberRepository memberRepository;
-    private final CommentDtoConverter commentDtoConverter;
+    private final ImageService imageService;
 
     @Transactional
     public CommentResponseDto createComment(Long postId, CommentRequestDto commentRequestDto, Long currentMemberId) {
@@ -39,24 +41,30 @@ public class CommentService {
         Comment comment = new Comment(commentRequestDto, member, post);
         Comment savedComment = commentRepository.save(comment);
 
-        return commentDtoConverter.toCommentResponseDto(savedComment, currentMemberId);
+        return CommentDtoConverter.toCommentResponseDto(savedComment, imageService.getFullImgUrl(member.getProfileImgUrl()), currentMemberId);
     }
 
     @Transactional
-    public CommentIdResponseDto updateComment(Long commentId, CommentRequestDto dto) {
+    public CommentIdResponseDto updateComment(Long postId, Long commentId, CommentRequestDto dto) {
         Comment comment = commentRepository.findById(commentId)
                 .orElseThrow(() -> new CustomException(ErrorCode.COMMENT_NOT_FOUND));
+        if (!comment.getPost().getPostId().equals(postId)) {
+            throw new CustomException(ErrorCode.COMMENT_NOT_FOUND);
+        }
 
         comment.updateContent(dto.getContent());
 
-        return commentDtoConverter.toCommentIdResponseDto(comment.getCommentId());
+        return CommentDtoConverter.toCommentIdResponseDto(comment.getCommentId());
     }
 
     @Transactional
-    public void deleteComment(Long commentId) {
+    public void deleteComment(Long postId, Long commentId) {
         Comment comment = commentRepository.findById(commentId)
                 .orElseThrow(() -> new CustomException(ErrorCode.COMMENT_NOT_FOUND));
 
+        if (!comment.getPost().getPostId().equals(postId)) {
+            throw new CustomException(ErrorCode.COMMENT_NOT_FOUND);
+        }
         commentRepository.delete(comment);
     }
 
@@ -64,7 +72,7 @@ public class CommentService {
         Post post = postRepository.findById(postId).orElseThrow(() -> new CustomException(ErrorCode.POST_NOT_FOUND));
         List<CommentResponseDto> commentList = new ArrayList<>();
         for (Comment c : post.getCommentList()){
-            commentList.add(commentDtoConverter.toCommentResponseDto(c, currentMemberId));
+            commentList.add(CommentDtoConverter.toCommentResponseDto(c, imageService.getFullImgUrl(c.getMember().getProfileImgUrl()), currentMemberId));
         }
         return commentList;
     }
